WIP: Fix invalid Cargo.toml keys
Its unclear if this is the right fix,
specifically with regards to enabling the features on tracing-futures.

Presumably this was working before, with the features disabled,
so its possible enalbing them may break things

Its also possible they were enabled elsewhere in the project,
which made Cargo enable them here too.